Copilot/career-application-copilot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.5
transformers==4.36.0
sentence-transformers==2.2.2
faiss-cpu>=1.9.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.1
spacy>=3.7.0
python-multipart
PyPDF2
python-docx
fuzzywuzzy
seaborn==0.13.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
textstat==0.7.3
matplotlib==3.7.2
# Remove duplicated packages
# textstat==0.7.3
# matplotlib==3.7.0

# Added dependencies for fine-tuned model
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.16.0
